refactor: extracted tab render functions to separate files
